pyproject.toml
[project]
name = "effective-mobile-test-task"
version = "0.1.0"
description = "A full-featured Django web application for managing cafe orders. The application allows you to add, delete, search, modify, and display orders."
authors = [
    {name = "Vladimir Volosnikov",email = "volosnikovvladimir69@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
force-exclude = '''
/(
    \.venv
  | \.venv_dev
  | migrations
)/
|
(
    .*/apps\.py
  | .*/settings\.py
  | manage\.py
  | /online_cafe/wsgi\.py
  | /online_cafe/asgi\.py
)
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
skip_glob = [
    "**/migrations/*",
    "**/settings.py",
    "**/apps.py",
    "**/.venv/*",
    "**/.venv_dev/*",
    "**wsgi.py",
    "**asgi.py",
    "**manage.py",
]
